backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
contextmemory>=0.1.0
openai>=1.57.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pycryptodome>=3.19.0
httpx>=0.26.0
email-validator>=2.0.0
